Review: Rename example to sample
