--- example/simple/app/simple/app-simple.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    alias(libs.plugins.fgp.app)
-}
-
-freeletics {
-    applicationId("com.freeletics.mad.example.simple")
-    useDaggerWithComponent()
-}
-
-dependencies {
-    implementation(libs.fl.navigator.runtime.compose)
-    implementation(projects.feature.a.implementation)
-    implementation(projects.feature.a.nav)
-    implementation(projects.feature.main)
-}
