--- example/simple/app/simple/src/main/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-        android:name="com.freeletics.mad.example.app.App"
-        android:icon="@mipmap/ic_launcher"
-        android:label="MAD Example">
-    </application>
-
-</manifest>
