--- example/simple/feature/a/nav/src/main/java/com/freeletics/mad/example/feature/a/nav/FeatureARoute.kt
@@ -1,7 +1,0 @@
-package com.freeletics.mad.example.feature.a.nav
-
-import com.freeletics.mad.navigator.NavRoot
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-object FeatureARoute : NavRoot
