--- example/simple/feature/main/src/main/java/com/freeletics/mad/example/feature/main/MainDI.kt
@@ -1,11 +1,0 @@
-package com.freeletics.mad.example.feature.main
-
-import com.freeletics.mad.navigator.compose.NavDestination
-import com.freeletics.mad.whetstone.AppScope
-import com.squareup.anvil.annotations.ContributesTo
-import dagger.Module
-import dagger.Provides
-
-@Module
-@ContributesTo(MainActivity::class)
-object MainRetainedModule
