+++ sample/simple/app/simple/app-simple.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    alias(libs.plugins.fgp.app)
+}
+
+freeletics {
+    applicationId("com.freeletics.mad.sample.simple")
+    useDaggerWithComponent()
+}
+
+dependencies {
+    implementation(libs.fl.navigator.runtime.compose)
+    implementation(projects.feature.a.implementation)
+    implementation(projects.feature.a.nav)
+    implementation(projects.feature.main)
+}
