+++ sample/simple/app/simple/src/main/java/com/freeletics/mad/sample/app/App.kt
@@ -1,0 +1,17 @@
+package com.freeletics.mad.sample.app
+
+import android.app.Application
+import com.freeletics.mad.whetstone.AppScope
+
+class App : Application() {
+    override fun getSystemService(name: String): Any? {
+        if (name == AppScope::class.qualifiedName) {
+            return component
+        }
+        return super.getSystemService(name)
+    }
+
+    private val component by lazy {
+        AppComponent.factory().create(this)
+    }
+}
