+++ sample/simple/app/simple/src/main/res/mipmap-anydpi/ic_launcher.xml
