+++ sample/simple/feature/a/implementation/feature-a-implementation.gradle.kts
