+++ sample/simple/feature/a/implementation/src/main/java/com/freeletics/mad/sample/feature/a/FeatureADI.kt
@@ -1,0 +1,9 @@
+package com.freeletics.mad.sample.feature.a
+
+import com.freeletics.mad.sample.feature.a.nav.FeatureARoute
+import com.squareup.anvil.annotations.ContributesTo
+import dagger.Module
+
+@Module
+@ContributesTo(FeatureARoute::class)
+object FeatureARetainedModule
