+++ sample/simple/feature/a/implementation/src/main/java/com/freeletics/mad/sample/feature/a/FeatureANavigator.kt
@@ -1,0 +1,11 @@
+package com.freeletics.mad.sample.feature.a
+
+import com.freeletics.mad.sample.feature.a.nav.FeatureARoute
+import com.freeletics.mad.navigator.NavEventNavigator
+import com.freeletics.mad.whetstone.ScopeTo
+import com.squareup.anvil.annotations.ContributesBinding
+import javax.inject.Inject
+
+@ScopeTo(FeatureARoute::class)
+@ContributesBinding(FeatureARoute::class, NavEventNavigator::class)
+class FeatureANavigator @Inject constructor() : NavEventNavigator()
