+++ sample/simple/feature/a/nav/feature-a-nav.gradle.kts
