+++ sample/simple/feature/a/nav/src/main/java/com/freeletics/mad/sample/feature/a/nav/FeatureARoute.kt
@@ -1,0 +1,7 @@
+package com.freeletics.mad.sample.feature.a.nav
+
+import com.freeletics.mad.navigator.NavRoot
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+object FeatureARoute : NavRoot
