+++ sample/simple/feature/main/feature-main.gradle.kts
