+++ sample/simple/feature/main/src/main/java/com/freeletics/mad/sample/feature/main/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.freeletics.mad.sample.feature.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.core.view.WindowCompat
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+
+        setContent {
+            WhetstoneMainScreen(Bundle.EMPTY)
+        }
+    }
+}
