+++ sample/simple/feature/main/src/main/java/com/freeletics/mad/sample/feature/main/MainDI.kt
@@ -1,0 +1,8 @@
+package com.freeletics.mad.sample.feature.main
+
+import com.squareup.anvil.annotations.ContributesTo
+import dagger.Module
+
+@Module
+@ContributesTo(MainActivity::class)
+object MainRetainedModule
