+++ sample/simple/gradle.properties
