+++ sample/simple/gradle/libs.versions.toml
