+++ sample/simple/gradlew.bat
