+++ sample/simple/settings.gradle.kts
@@ -1,0 +1,13 @@
+pluginManagement {
+    repositories {
+        google()
+        gradlePluginPortal()
+        mavenCentral()
+    }
+}
+
+plugins {
+    id("com.freeletics.gradle.settings").version("0.2.0")
+}
+
+rootProject.name = "simple-sample"
